+++ client/src/pages/student/Dashboard.jsx
@@ -1,9 +1,10 @@
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/layout/Navbar";
 import TaskList from "../sections/TaskList";
 import ClassSched from "../sections/ClassSched";
 import ClassResources from "../sections/ClassResources";
 import ActivityTracker from "../sections/ActivityTracker";
 import ClassList from "../sections/ClassList";
+import Footer from "../../components/layout/Footer";
 
 function Dashboard() {
   return (
@@ -15,9 +16,8 @@ function Dashboard() {
         <TaskList />
         <ClassSched />
         <ClassResources />
-        <ActivityTracker />
-        <ClassList />
       </div>
+      <Footer />
     </div>
   );
 }