+++ client/src/pages/student/Register.jsx
@@ -2,10 +2,10 @@ import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 import { registerAccount, getMe } from "../../services/auth";
 import { Icon } from "@iconify/react";
-import Button from "../../components/Button";
-import Modal from "../../components/Modal";
+import Button from "../../components/ui/Button";
+import Modal from "../../components/ui/Modal";
 import { handleApiError } from "../../services/errorHandler";
-import LoadingScreen from "../../components/LoadingScreen";
+import LoadingScreen from "../../components/ui/LoadingScreen";
 
 function Register(){
 	const isNumber = (value) => /^[0-9]+$/.test(value);
@@ -95,7 +95,7 @@ function Register(){
 			<div className="absolute inset-0 flex items-center justify-center z-1 px-4">
 				<form onSubmit={handleRegister} className="bg-[#FAF9F6] flex flex-col p-5 rounded-xl shadow-[0_8px_30px_rgba(0,0,0,0.12)] w-80 max-w sm:max-w-sm">
 					<div className="flex items-center justify-center mb-7 z-10">
-						<img className="w-12" src="/assets/Kabsupanion-Logo.png" alt="Logo"/>
+						<img className="w-12" src="/assets/images/Kabsupanion-Logo.png" alt="Logo"/>
 						<p className="font-bold text-2xl pl-2 text-[#1B651B] font-['Roboto_Condensed']">Registration Form</p>
 					</div>
           
@@ -144,7 +144,7 @@ function Register(){
 
 			{/* LAYA AT DIWA BACKGROUND */}
 			<div className="sm:block absolute bottom-0 right-4 md:right-20 lg:right-80 z-0">
-				<img className="opacity-50 w-72" src="/assets/Laya-at-Diwa.png" alt="Laya at Diwa"/>
+				<img className="opacity-50 w-72" src="/assets/images/Laya-at-Diwa.png" alt="Laya at Diwa"/>
 			</div>
 		</div>
   );